provider_backend/requirements.test.txt
aiojobs==0.3.0
aiohttp==3.6.3
aiomysql==0.0.21
aiopg==1.3.2
aioprometheus[aiohttp,binary]
cryptography
jsonschema==3.2.0
mysql-connector-python
psycopg2-binary
pyhamcrest
pytest
pytest-ordering
pytest-asyncio
redis
requests